+++ R/AAA.R
@@ -4,7 +4,7 @@
 #' @importFrom purrr flatten flatten_chr imap
 #' @importFrom withr with_environment
 #' @importFrom assertthat assert_that
-#' @importFrom ggplot2 geom_line geom_point ggplot aes_string xlab ylab
+#' @importFrom ggplot2 geom_line geom_point ggplot xlab ylab
 #' @importFrom ggplot2 guides theme_bw scale_y_continuous scale_y_log10
 #' @importFrom ggplot2 facet_grid facet_wrap aes scale_color_discrete theme
 #' @importFrom ggplot2 scale_color_viridis_c scale_color_brewer labs